+++ lib/extended_logger.dart
@@ -1,5 +1,5 @@
 export 'extended_logger.dart';
-export 'src/___logger.dart';
 export 'src/config.dart';
 export 'src/google_cloud_run_utils.dart';
 export 'src/log_level.dart';
+export 'src/logger.dart';